requirements.txt
streamlit==1.32.0
google-generativeai==0.3.2
requests==2.31.0
beautifulsoup4==4.12.3
PyPDF2==3.0.1